src/common/requirements.txt
# 共通ライブラリの依存パッケージ
# バージョン範囲は上限を設定し、破壊的変更を防止

fastapi>=0.100.0,<0.110.0       # 0.110.0はPydantic v3必須の破壊的変更あり
pydantic[email]>=2.0.0,<3.0.0   # v2系のみサポート、v3は大幅なAPI変更
python-jose[cryptography]>=3.3.0,<4.0.0  # v4でAPI変更の可能性
passlib[bcrypt]>=1.7.4,<2.0.0   # bcrypt互換性維持
azure-cosmos>=4.5.0,<5.0.0      # v5で非同期APIが大幅変更
python-multipart>=0.0.6,<1.0.0  # メジャーバージョンアップ対策
httpx>=0.24.0,<1.0.0            # 非同期HTTPクライアント
tenacity>=8.2.0,<9.0.0          # リトライ機能
aiohttp>=3.8.0,<4.0.0           # Cosmos DB非同期トランスポート

# 推奨バージョン（開発・テストで使用）
# fastapi==0.109.0
# pydantic[email]==2.5.3
# python-jose[cryptography]==3.3.0
# passlib[bcrypt]==1.7.4
# azure-cosmos==4.5.1
# python-multipart==0.0.6
# httpx==0.24.1
# tenacity==8.2.3
